Fix import paths and refactor context hooks to use separate hook files
